+++ src/main/java/com/ninjaone/dundie_awards/controller/IndexController.java
@@ -2,8 +2,8 @@ package com.ninjaone.dundie_awards.controller;
 
 import com.ninjaone.dundie_awards.AwardsCache;
 import com.ninjaone.dundie_awards.MessageBroker;
+import com.ninjaone.dundie_awards.employee.EmployeeRepository;
 import com.ninjaone.dundie_awards.repository.ActivityRepository;
-import com.ninjaone.dundie_awards.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
