+++ src/main/java/com/ninjaone/dundie_awards/employee/EmployeeRepository.java
@@ -1,6 +1,5 @@
-package com.ninjaone.dundie_awards.repository;
+package com.ninjaone.dundie_awards.employee;
 
-import com.ninjaone.dundie_awards.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
